requirements.txt
tensorflow>=2.10
torch>=1.12
torchvision>=0.13
opencv-python
numpy
matplotlib
scikit-learn
seaborn
pillow
tqdm